# Speed up method `ResizeGenerator.forward` by 131%