--- a/kornia/geometry/transform/affwarp.py
+++ b/kornia/geometry/transform/affwarp.py
@@ -27,7 +27,8 @@ from kornia.utils import _extract_device_dtype
 from kornia.utils.image import perform_keep_shape_image
 from kornia.utils.misc import eye_like
 
-from .imgwarp import get_affine_matrix2d, get_projective_transform, get_rotation_matrix2d, warp_affine, warp_affine3d
+from .imgwarp import (get_affine_matrix2d, get_projective_transform,
+                      get_rotation_matrix2d, warp_affine, warp_affine3d)
 
 __all__ = [
     "Affine",