<!-- CODEFLASH_OPTIMIZATION: {"function":"ResizeGenerator.forward","file":"kornia/augmentation/random_generator/_2d/resize.py","speedup_pct":"131%","speedup_x":"1.31x","original_runtime":"2.10 milliseconds","best_runtime":"913 microseconds","optimization_type":"memory","timestamp":"2025-08-07T23:19:03.065Z","version":"1.0"} -->
### 📄 131% (1.31x) speedup for ***`ResizeGenerator.forward` in `kornia/augmentation/random_generator/_2d/resize.py`***

⏱️ Runtime :   **`2.10 milliseconds`**  **→** **`913 microseconds`** (best of `672` runs)
### 📝 Explanation and details

Here is an optimized version of the provided program for runtime and memory efficiency, preserving function signatures and return values. Comments are left unchanged, except where code was modified to optimize logic.

**Key optimizations:**
- Eliminate intermediate tensors where possible (e.g., use built-in tensor arithmetic for bbox generation instead of many `repeat` and in-place operations).
- Remove repeated construction of 0 tensors and batch/reshape in bbox_generator to use less memory and improve PyTorch backend parallelism.
- Minimize scalar tensor creation in `forward` (replace `tensor(x, ...)` with `torch.full` or simple operations for batch tensors).
- Inline single-use variables and clarify usage of batch primitives for vectorization, for more efficient backwards and reduced memory pressure.



**Notes:**
- The biggest speedup comes from vectorizing `bbox_generator` and avoiding repeated/scalar tensor construction. This avoids unnecessary data movement and leverages PyTorch's core routines.
- The new implementation yields the exact same output shapes and numerical results as the original, but performs better for large batches and doesn't allocate or repeat constants inefficiently.



✅ **Correctness verification report:**

| Test                        | Status            |
| --------------------------- | ----------------- |
| ⚙️ Existing Unit Tests | 🔘 **None Found** |
| 🌀 Generated Regression Tests | ✅ **40 Passed** |
| ⏪ Replay Tests | 🔘 **None Found** |
| 🔎 Concolic Coverage Tests | 🔘 **None Found** |
|📊 Tests Coverage       | 100.0% |
<details>
<summary>🌀 Generated Regression Tests and Runtime</summary>

```python
from __future__ import annotations

from typing import Dict, Optional, Tuple, Union

# imports
import pytest  # used for our unit tests
import torch
from kornia.augmentation.random_generator._2d.resize import ResizeGenerator


class RandomGeneratorBase(torch.nn.Module):
    """Base class for generating random augmentation parameters."""
    device: Optional[torch.device] = None
    dtype: torch.dtype = torch.float32

    def __init__(self) -> None:
        super().__init__()

    def set_rng_device_and_dtype(self, device: Optional[torch.device] = None, dtype: Optional[torch.dtype] = None):
        if device is not None:
            self.device = device
        if dtype is not None:
            self.dtype = dtype
from kornia.augmentation.random_generator._2d.resize import ResizeGenerator

# unit tests

# -------------------
# BASIC TEST CASES
# -------------------

def test_forward_basic_tuple_resize():
    # Test basic resize with tuple, batch_size=1, shape (1, 3, 32, 64) -> resize to (16, 32)
    gen = ResizeGenerator(resize_to=(16, 32))
    batch_shape = (1, 3, 32, 64)
    codeflash_output = gen.forward(batch_shape); out = codeflash_output # 144μs -> 60.3μs (139% faster)
    # Check src bbox corners
    expected_src = torch.tensor([[[0, 0], [63, 0], [63, 31], [0, 31]]], dtype=out["src"].dtype)
    # Check dst bbox corners
    expected_dst = torch.tensor([[[0, 0], [31, 0], [31, 15], [0, 15]]], dtype=out["dst"].dtype)

def test_forward_basic_int_resize_short_side():
    # Test resize with int, side="short", batch_size=2, shape (2, 3, 20, 40)
    gen = ResizeGenerator(resize_to=10, side="short")
    batch_shape = (2, 3, 20, 40)
    codeflash_output = gen.forward(batch_shape); out = codeflash_output # 146μs -> 61.8μs (137% faster)
    # Check dst bbox
    expected_dst = torch.tensor([[[0, 0], [19, 0], [19, 9], [0, 9]],
                                [[0, 0], [19, 0], [19, 9], [0, 9]]], dtype=out["dst"].dtype)

def test_forward_basic_int_resize_horz_side():
    # Test resize with int, side="horz", batch_size=1, shape (1, 3, 30, 60)
    gen = ResizeGenerator(resize_to=15, side="horz")
    batch_shape = (1, 3, 30, 60)
    codeflash_output = gen.forward(batch_shape); out = codeflash_output # 143μs -> 59.7μs (140% faster)
    expected_dst = torch.tensor([[[0, 0], [14, 0], [14, 6], [0, 6]]], dtype=out["dst"].dtype)

def test_forward_basic_batch_size_3():
    # Test batch_size=3, tuple resize
    gen = ResizeGenerator(resize_to=(8, 8))
    batch_shape = (3, 1, 8, 8)
    codeflash_output = gen.forward(batch_shape); out = codeflash_output # 144μs -> 60.3μs (139% faster)

def test_forward_basic_dtype_and_device():
    # Test with float64 and cuda if available
    gen = ResizeGenerator(resize_to=(6, 6))
    device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
    gen.set_rng_device_and_dtype(device=device, dtype=torch.float64)
    batch_shape = (1, 1, 6, 6)
    codeflash_output = gen.forward(batch_shape); out = codeflash_output # 144μs -> 59.7μs (143% faster)

# -------------------
# EDGE CASES
# -------------------

def test_forward_zero_batch():
    # batch_size=0 should return empty tensors of correct shape
    gen = ResizeGenerator(resize_to=(5, 5))
    batch_shape = (0, 3, 10, 10)
    codeflash_output = gen.forward(batch_shape); out = codeflash_output # 5.79μs -> 5.08μs (13.9% faster)

def test_forward_invalid_batch_size_negative():
    # batch_size < 0 should raise AssertionError
    gen = ResizeGenerator(resize_to=(5, 5))
    with pytest.raises(AssertionError):
        gen.forward((-1, 3, 10, 10)) # 875ns -> 833ns (5.04% faster)

def test_forward_invalid_resize_to_tuple():
    # resize_to not positive ints or wrong length should raise
    gen = ResizeGenerator(resize_to=(0, 5))
    with pytest.raises(AssertionError):
        gen.forward((1, 1, 10, 10)) # 76.6μs -> 30.2μs (154% faster)
    gen = ResizeGenerator(resize_to=(5, -2))
    with pytest.raises(AssertionError):
        gen.forward((1, 1, 10, 10)) # 70.2μs -> 26.7μs (163% faster)
    gen = ResizeGenerator(resize_to=(5, 5, 5))
    with pytest.raises(AssertionError):
        gen.forward((1, 1, 10, 10)) # 69.0μs -> 26.0μs (166% faster)
    # resize_to with float values
    gen = ResizeGenerator(resize_to=(5.5, 5))
    with pytest.raises(AssertionError):
        gen.forward((1, 1, 10, 10)) # 68.8μs -> 26.3μs (162% faster)

def test_forward_invalid_side():
    # Invalid side string should raise ValueError
    gen = ResizeGenerator(resize_to=10, side="invalid")
    with pytest.raises(ValueError):
        gen.forward((1, 1, 10, 10)) # 73.9μs -> 28.6μs (158% faster)

def test_forward_same_on_batch_non_bool():
    # same_on_batch should be boolean or None
    gen = ResizeGenerator(resize_to=(5, 5))
    with pytest.raises(AssertionError):
        gen.forward((1, 1, 10, 10), same_on_batch="yes") # 1.00μs -> 1.04μs (4.03% slower)

def test_forward_minimal_input_size():
    # Minimal input size (1, 1, 1, 1)
    gen = ResizeGenerator(resize_to=(1, 1))
    codeflash_output = gen.forward((1, 1, 1, 1)); out = codeflash_output # 146μs -> 61.5μs (139% faster)

# -------------------
# LARGE SCALE TEST CASES
# -------------------

def test_forward_large_batch():
    # Large batch size, but <1000 elements
    gen = ResizeGenerator(resize_to=(20, 30))
    batch_shape = (999, 3, 50, 60)
    codeflash_output = gen.forward(batch_shape); out = codeflash_output # 148μs -> 95.6μs (55.8% faster)

def test_forward_large_image_size():
    # Large image size, but <100MB
    # Each bbox is (batch, 4, 2), so for batch=1, 4*2*8=64 bytes (float64)
    # Let's use batch=2, image 512x512, output 256x256
    gen = ResizeGenerator(resize_to=(256, 256))
    gen.set_rng_device_and_dtype(dtype=torch.float64)
    batch_shape = (2, 3, 512, 512)
    codeflash_output = gen.forward(batch_shape); out = codeflash_output # 146μs -> 62.4μs (135% faster)

def test_forward_large_batch_and_image():
    # batch=100, image 128x256, output 64x128
    gen = ResizeGenerator(resize_to=(64, 128))
    batch_shape = (100, 3, 128, 256)
    codeflash_output = gen.forward(batch_shape); out = codeflash_output # 145μs -> 65.5μs (123% faster)

def test_forward_large_varied_aspect_ratio():
    # batch=10, image 100x10 (aspect ratio 10), resize_to=5, side="short"
    gen = ResizeGenerator(resize_to=5, side="short")
    batch_shape = (10, 3, 10, 100)
    codeflash_output = gen.forward(batch_shape); out = codeflash_output # 145μs -> 62.8μs (132% faster)

# -------------------
# MUTATION TESTING: SENSITIVITY TO LOGIC
# -------------------

def test_forward_mutation_wrong_src_bbox():
    # If the src bbox is not calculated as expected, this should fail
    gen = ResizeGenerator(resize_to=(4, 8))
    batch_shape = (1, 3, 4, 8)
    codeflash_output = gen.forward(batch_shape); out = codeflash_output # 141μs -> 59.9μs (137% faster)
    # The src bbox should be [[0,0],[7,0],[7,3],[0,3]]
    expected_src = torch.tensor([[[0, 0], [7, 0], [7, 3], [0, 3]]], dtype=out["src"].dtype)

def test_forward_mutation_wrong_dst_bbox():
    # If the dst bbox is not calculated as expected, this should fail
    gen = ResizeGenerator(resize_to=(2, 6))
    batch_shape = (1, 3, 4, 8)
    codeflash_output = gen.forward(batch_shape); out = codeflash_output # 140μs -> 58.7μs (140% faster)
    expected_dst = torch.tensor([[[0, 0], [5, 0], [5, 1], [0, 1]]], dtype=out["dst"].dtype)
# codeflash_output is used to check that the output of the original code is the same as that of the optimized code.
```

</details>